Add controller test
